refactor:change import paths in auth card and sign up
