+++ src/features/auth/ui/authCard/authCard.tsx
@@ -1,6 +1,6 @@
 import { ComponentPropsWithoutRef } from 'react'
 
-import { GithubSvgrepoCom31, GoogleSvgrepoCom1 } from '@/assets'
+import { GithubSvgrepoCom31, GoogleSvgrepoCom1 } from '@/assets/icons'
 import { Button, Card, Typography } from '@photo-fiesta/ui-lib'
 import Link from 'next/link'
