+++ src/features/auth/ui/signIn/useSignIn.ts
@@ -1,19 +1,17 @@
 import { SubmitHandler, useForm } from 'react-hook-form'
 
+import { PASSWORD_REGEX } from '@/shared/config/regex-constants'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
 
 export type FormInputs = z.infer<typeof signInSchema>
-{
-  /**TODO: use global variable regex*/
-}
 const signInSchema = z.object({
   email: z.string().email('Invalid email address'),
   password: z
     .string()
     .min(6, 'Password must be at least 8 characters')
     .max(20, 'Password must be at most 32 characters')
-    .regex(/^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*()_+\-=\\[\]{};':"\\|,.<>/?]).*$/),
+    .regex(PASSWORD_REGEX),
 })
 
 export const useSignIn = (onSubmit: (data: FormInputs) => void) => {
