+++ src/pages/index.tsx
@@ -1,7 +1,6 @@
-import { SentEmail } from '@/features/auth/ui'
+import { ConfirmEmail, SentEmail, SignIn } from '@/features/auth/ui'
 import { SignUp } from '@/features/auth/ui/signUp'
 import { Modal } from '@photo-fiesta/ui-lib'
-
 
 export default function Home() {
   function handleSubmit(data: unknown) {
@@ -16,6 +15,7 @@ export default function Home() {
       <Modal />
       <ConfirmEmail />
       <SentEmail />
+      <SignIn onSubmit={handleSubmit} />
     </>
   )
 }
